Implement data persistence
A new scene startMenu was added to the build.  A start, exit and player name field were added for UI.

Player name from the field will persist between the game and the menu.

High scores will be saved between scenes and if the application is exited using the exit button, the high scores will save to file for the next session.